pyproject.toml
[project]
name = "mlpipeline"
version = "0.1.0"
description = "ML pipeline migration showcase"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "lightgbm",
  "scikit-learn",
  "mlflow",
  "duckdb",
  "pandas",
  "numpy",
  "pyyaml",
  "ray",
]

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
